requirements.txt
streamlit
pandas
python-dotenv
httpx